Fix format-check issues for automatic compilation

+++ MosaicSrc/trgrecorderparser.cpp
@@ -97,7 +97,7 @@ long TrgRecorderParser::parse(int numClosed)
     trgNum  = buf2uint32(p);
     trgTime = buf2uint64(p + 4);
 
-    if (verbose) printf("Trigger %d @ %ld\n", trgNum, trgTime);
+    if (verbose) printf("Trigger %d @ %llu\n", trgNum, static_cast<unsigned long long>(trgTime));
 
     p += evSize;
     numClosed--;
